Refactored the Content Page
I split up the code from the main content page and created 3 sub files (Trackspage, CharactersPagee, and ItemsPage). I also moved reusable code into the components folder and created a new data folder to hold long data lists to improve readability in page files/
