+++ src/App.tsx
@@ -1,33 +1,26 @@
 /*import { useState } from 'react';*/
-import {
-  BrowserRouter as Router,
-  Routes,
-  Route,
-} from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
-import LoginPage from './pages/Login';
-import SignupPage from './pages/Signup';
-import AboutUsPage from './pages/AboutUs';
-import ProcessPage from './pages/Process';
-import {
-  ContentPage,
-  CharactersPage,
-  ItemsPage,
-  TracksPage,
-}
-  from './pages/Content';
-import ForumPage from './pages/Forum';
-import HomePage from './pages/Home';
-import LeaderboardPage from './pages/Leaderboard';
-import NewsAndUpdatesPage from './pages/NewsAndUpdates';
-import PlayerStatsPage from './pages/PlayerStats';
-import NoPage from './pages/NoPage';
-import Navbar from './Navbar';
-import ForumPost from './pages/ForumPost';
-import Footer from './components/footer';
-import { useEffect, useState } from 'react';
+import LoginPage from "./pages/Login";
+import SignupPage from "./pages/Signup";
+import AboutUsPage from "./pages/AboutUs";
+import ProcessPage from "./pages/Process";
+import { ContentPage } from "./pages/content/Content";
+import { TracksPage } from "./pages/content/TracksPage";
+import { CharactersPage } from "./pages/content/CharactersPage";
+import { ItemsPage } from "./pages/content/ItemsPage";
+import ForumPage from "./pages/Forum";
+import HomePage from "./pages/Home";
+import LeaderboardPage from "./pages/Leaderboard";
+import NewsAndUpdatesPage from "./pages/NewsAndUpdates";
+import PlayerStatsPage from "./pages/PlayerStats";
+import NoPage from "./pages/NoPage";
+import Navbar from "./Navbar";
+import ForumPost from "./pages/ForumPost";
+import Footer from "./components/footer";
+import { useEffect, useState } from "react";
 
-export interface AccountSchema { 
+export interface AccountSchema {
   pid: string;
   username: string;
 }
@@ -39,42 +32,60 @@ function App() {
   useEffect(() => {
     const storedPID: string | null = localStorage.getItem("pid");
     const storedUsername: string | null = localStorage.getItem("username");
-    if (!storedPID || !storedUsername) { setAccount(null); return; }
+    if (!storedPID || !storedUsername) {
+      setAccount(null);
+      return;
+    }
     const storedAccount: AccountSchema = {
       pid: storedPID,
-      username: storedUsername
-    }
+      username: storedUsername,
+    };
     setAccount(storedAccount);
   }, []);
 
   return (
     <Router basename="/ggk">
-      <Navbar account={account} setAccount={setAccount} setCurrentPage={setCurrentPage} currentPage={currentPage}/>
+      <Navbar
+        account={account}
+        setAccount={setAccount}
+        setCurrentPage={setCurrentPage}
+        currentPage={currentPage}
+      />
 
-      <div className='min-h-[80vh]'>
+      <div className="min-h-[80vh]">
         <Routes>
-        <Route path="/login" element={<LoginPage setAccount={setAccount}/>} />
-        <Route path="/signup" element={<SignupPage setAccount={setAccount}/>} />
-        <Route path="/aboutUs" element={<AboutUsPage />} />
-        <Route path="/process" element={<ProcessPage/>}/>
-        <Route path="/content" element={<ContentPage />} />
-        <Route path="/forum" element={<ForumPage />} />
-        <Route path="/home" element={<HomePage setCurrentPage={setCurrentPage}/>} />
-        <Route path="/leaderboard" element={<LeaderboardPage />} />
-        <Route path="/news" element={<NewsAndUpdatesPage />} />
-        <Route path="/stats" element={<PlayerStatsPage account={account}/>} />
-        <Route path="/forumPost" element={<ForumPost />} />
-        <Route path="/characters" element={<CharactersPage />} />
-        <Route path="/tracks" element={<TracksPage />} />
-        <Route path="/items" element={<ItemsPage />} />
-        <Route path="*" element={<NoPage/>}></Route>
-      </Routes>
+          <Route
+            path="/login"
+            element={<LoginPage setAccount={setAccount} />}
+          />
+          <Route
+            path="/signup"
+            element={<SignupPage setAccount={setAccount} />}
+          />
+          <Route path="/aboutUs" element={<AboutUsPage />} />
+          <Route path="/process" element={<ProcessPage />} />
+          <Route path="/content" element={<ContentPage />} />
+          <Route path="/forum" element={<ForumPage />} />
+          <Route
+            path="/home"
+            element={<HomePage setCurrentPage={setCurrentPage} />}
+          />
+          <Route path="/leaderboard" element={<LeaderboardPage />} />
+          <Route path="/news" element={<NewsAndUpdatesPage />} />
+          <Route
+            path="/stats"
+            element={<PlayerStatsPage account={account} />}
+          />
+          <Route path="/forumPost" element={<ForumPost />} />
+          <Route path="/characters" element={<CharactersPage />} />
+          <Route path="/tracks" element={<TracksPage />} />
+          <Route path="/items" element={<ItemsPage />} />
+          <Route path="*" element={<NoPage />}></Route>
+        </Routes>
       </div>
       <Footer />
     </Router>
-  
   );
 }
 
 export default App;
-
