+++ src/data/characters.ts
@@ -1,0 +1,83 @@
+
+export interface Character {
+    imgUrl: string;
+    name: string;
+    description: string;
+    songName: string;
+    songLink: string;
+}
+
+
+
+
+// sample list of characters to display
+export const characters: Character[] = [
+    {
+      imgUrl: "images/characters/FrshSkater.PNG",
+      name: "Morgan",
+      description: `Pronouns: She/They
+      Year: Freshmen
+      Major: Photography
+      Likes: Skating, Thrifting Clothes, Going Fast, Finding Spots
+      Dislikes: Waking Up For Class`,
+      songName: "Skater Boi - Avril Lavigne",
+      songLink: "https://www.youtube.com/watch?v=TIy3n2b7V9k",
+    },
+    {
+      imgUrl: "images/characters/SophDining.png",
+      name: "Reese",
+      description: `Pronouns: He/Him
+      Year: Sophomore
+      Major: Packaging Science
+      Likes: Free Stuff, Creating Playlists For Events, Anime, Energy Drinks
+      Dislikes: Working 3 Jobs, Sleep`,
+      songName: "Feel Good Inc. - Gorillaz",
+      songLink: "https://www.youtube.com/watch?v=HyHNuVaZJ-k",
+    },
+    {
+      imgUrl: "images/characters/OLjr.png",
+      name: "Emma",
+      description: `Pronouns: She/Her
+      Year: Junior
+      Job: Orientation Leader
+      Major: Game Design and Development
+      Likes: Pins, Walking, Early Morning Walks, Bucket Hats
+      Dislikes: Loud Music, Sleeping In`,
+      songName: "Turbo Hustle - DJ Maestro",
+      songLink:
+        "https://www.youtube.com/watch?v=LGpaGI99Xl0&list=RDLGpaGI99Xl0&start_radio=1",
+    },
+    {
+      imgUrl: "images/characters/HkySr.png",
+      name: "Kai",
+      description: `Pronouns: They/Them
+      Major: Business
+      Likes: Hockey, Bandanas, Going On Adventures, Being Team Captain
+      Dislikes: Sitting around, Losing a game`,
+      songName: `Eye of the Tiger - Survivor`,
+      songLink:
+        "https://www.youtube.com/watch?v=btPJPFnesV4&list=RDbtPJPFnesV4&start_radio=1",
+    },
+    {
+      imgUrl: "images/characters/Jamster.png",
+      name: "Jamster",
+      description: `Pronouns: They/Them
+      Year: Primordior
+      Major: All
+      Likes: Gaming, Jamming, Running Around, GameJams
+      Dislikes: Not Getting Snacks`,
+      songName: "Hamster Dance -  Hampton and the Hamsters",
+      songLink: "https://www.youtube.com/watch?v=p3G5IXn0K7A",
+    },
+    {
+      imgUrl: "images/characters/gizmo.jpeg",
+      name: "Gizmo",
+      description: `Pronouns: He/Him
+      Year: Sophomore
+      Major: Packaging Science
+      Likes: Free Stuff, Creating Playlists For Events, Anime, Energy Drinks
+      Dislikes: Working 3 Jobs, Sleep`,
+      songName: "Feel Good Inc. - Gorillaz",
+      songLink: "https://www.youtube.com/watch?v=HyHNuVaZJ-k",
+    },
+  ];
